refactor out default variables and isSelectedVariablesUpdated into separate util
