+++ ui/dashboards/src/utils/index.ts
@@ -13,3 +13,4 @@
 
 export * from './panelUtils';
 export * from './time';
+export * from './variableUtils';
